refactor: 원시값 포장

+++ src/main/kotlin/lotto/domain/LottoResults.kt
@@ -14,7 +14,7 @@ data class LottoResults(
     }
 
     private fun sumOfResult(): Int = lottoResults.sumOf {
-        it.lottoPrize.prizeMoney * it.matchedLottoCount
+        it.lottoPrize.prizeMoney * it.matchedLottoCount.toInt()
     }
 }
